+++ R/vec-na.R
@@ -22,8 +22,6 @@
 #' contain any types, expressions like `list(NA)` store a logical
 #' `NA`.
 #'
-#' @seealso The [new-vector-along] family to create typed vectors filled
-#'   with missing values.
 #' @examples
 #' typeof(NA)
 #' typeof(na_lgl)